migrated from internal computation from double to Bigdecimal to avoid floating point rounding errors

+++ src/test/java/de/lgohlke/codedemo/service/StatisticBucketPerSecondTest.java
@@ -24,6 +24,6 @@ public class StatisticBucketPerSecondTest {
         bucket.resetWhenExpired(11, 10);
 
         assertThat(bucket.getCount()).isEqualTo(1);
-        assertThat(bucket.getSum()).isEqualTo(10);
+        assertThat(bucket.getSum().doubleValue()).isEqualTo(10);
     }
 }
